tool/requirements.txt
openpyxl~=3.1.0
selenium~=4.8.0
flask
requests
mysql-connector-python
